CLEANUP | Extract functionality from main component
The main app component was getting a bit crowded for my taste. This
commit

* Moves the actual generation logic into it's own file
* Moves the preset list into it's own component

Now the main app component is a bit leaner again.
